+++ src/components/select/simple-select/simple-select.spec.js
@@ -10,7 +10,6 @@ import SimpleSelect from "./simple-select.component";
 import Textbox from "../../textbox";
 import Option from "../option/option.component";
 import SelectList from "../select-list/select-list.component";
-import SelectTextbox from "../select-textbox/select-textbox.component";
 import InputIconToggleStyle from "../../../__internal__/input-icon-toggle/input-icon-toggle.style";
 import InputPresentationStyle from "../../../__internal__/input/input-presentation.style";
 import Label from "../../../__internal__/label";
@@ -61,24 +60,11 @@ describe("SimpleSelect", () => {
   });
 
   describe("disablePortal", () => {
-    it("renders SelectList as a content of positionedChildren prop on Textbox when disablePortal is true", () => {
+    it("renders SelectList with a disablePortal prop assigned", () => {
       const wrapper = renderSelect({ disablePortal: true });
 
-      simulateSelectTextEvent(wrapper, "click");
-      const positionedChildren = mount(
-        wrapper.find(SelectTextbox).props().positionedChildren
-      );
-      expect(positionedChildren.find(SelectList).exists()).toBe(true);
-    });
-
-    it("renders SelectList as a direct children of StyledSimpleSelect by default", () => {
-      const wrapper = renderSelect();
-
-      simulateSelectTextEvent(wrapper, "click");
-      expect(wrapper.find(SelectTextbox).props().positionedChildren).toBe(
-        false
-      );
-      expect(wrapper.find(SelectList).exists()).toBe(true);
+      wrapper.find(Textbox).find('[type="dropdown"]').first().simulate("click");
+      expect(wrapper.find(SelectList).props().disablePortal).toBe(true);
     });
   });